requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.0
rich==13.7.1
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
itsdangerous==2.2.0
starlette==0.38.6
